# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeScope API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user:
    post:
      operationId: upsertUser
      tags: [user]
      summary: Create or update user from Telegram data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: User upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /user/{telegramId}:
    get:
      operationId: getUser
      tags: [user]
      summary: Get user profile and stats
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /user/{telegramId}/stats:
    get:
      operationId: getUserStats
      tags: [user]
      summary: Get trading DNA and analytics for a user
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User stats and trading DNA
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
  /user/{telegramId}/weekly-report:
    get:
      operationId: getWeeklyReport
      tags: [user]
      summary: Get AI-generated weekly performance report
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
...
  /risk/calculate:
    post:
      operationId: calculateRisk
      tags: [analyze]
      summary: Advanced risk and position size calculator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskCalcInput"
      responses:
        "200":
          description: Risk calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskCalcResult"
  /risk/scenarios/{userId}:
    get:
      operationId: getRiskScenarios
      tags: [analyze]
      summary: Get saved risk scenarios for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Risk scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskScenario"
    post:
      operationId: saveRiskScenario
      tags: [analyze]
      summary: Save a risk scenario
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskScenarioInput"
      responses:
        "201":
          description: Scenario saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskScenario"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserInput:
      type: object
      required: [telegramId]
      properties:
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    User:
      type: object
      required: [id, telegramId, currentStreak, longestStreak, subscriptionTier, createdAt]
      properties:
        id:
          type: string
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        accountSize:
          type: number
        defaultRisk:
          type: number
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        subscriptionTier:
          type: string
        createdAt:
          type: string
    UserProfile:
      type: object
      required: [user, recentAnalyses, quickStats]
      properties:
        user:
          $ref: "#/components/schemas/User"
        recentAnalyses:
          type: array
          items:
            $ref: "#/components/schemas/AnalysisSummary"
        quickStats:
          $ref: "#/components/schemas/QuickStats"
    QuickStats:
      type: object
      required: [totalAnalyses, avgScoreThisWeek, bestScoreEver, winRate]
      properties:
        totalAnalyses:
          type: integer
        avgScoreThisWeek:
          type: number
        bestScoreEver:
          type: integer
        winRate:
          type: number
    UserStats:
      type: object
      required: [tradingDna, streaks, personalStats, mistakeHeatmap]
      properties:
        tradingDna:
          $ref: "#/components/schemas/TradingDna"
        streaks:
          $ref: "#/components/schemas/Streaks"
        personalStats:
          $ref: "#/components/schemas/PersonalStats"
        mistakeHeatmap:
          type: array
          items:
            $ref: "#/components/schemas/HeatmapCell"
        scoreHistory:
          type: array
          items:
            $ref: "#/components/schemas/ScoreHistoryPoint"
    TradingDna:
      type: object
      properties:
        strongestPair:
          type: ["string", "null"]
        bestTimeframe:
          type: ["string", "null"]
        bestSession:
          type: ["string", "null"]
        worstHabit:
          type: ["string", "null"]
        consistencyGrade:
          type: ["string", "null"]
        unlocked:
          type: boolean
    Streaks:
      type: object
      required: [currentStreak, longestStreak, disciplineStreak, goodRRStreak, newsSafeStreak]
      properties:
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        disciplineStreak:
          type: integer
        goodRRStreak:
          type: integer
        newsSafeStreak:
          type: integer
    PersonalStats:
      type: object
      required: [totalAnalyses, avgScoreAllTime, avgScoreThisWeek, avgScoreLastWeek, mostAnalyzedPair, winRate]
      properties:
        totalAnalyses:
          type: integer
        avgScoreAllTime:
          type: number
        avgScoreThisWeek:
          type: number
        avgScoreLastWeek:
          type: number
        mostAnalyzedPair:
          type: ["string", "null"]
        winRate:
          type: number
    HeatmapCell:
      type: object
      required: [day, hour, avgScore, count]
      properties:
        day:
          type: integer
        hour:
          type: integer
        avgScore:
          type: number
        count:
          type: integer
    ScoreHistoryPoint:
      type: object
      required: [date, score]
      properties:
        date:
          type: string
        score:
          type: number
    WeeklyReport:
      type: object
      required: [summary, strengths, improvements, tip, generatedAt]
      properties:
        summary:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        tip:
          type: string
        generatedAt:
          type: string
...
    AnalysisSummary:
      type: object
      required: [id, pair, direction, overallScore, scoreLabel, outcome, createdAt]
      properties:
        id:
          type: string
        pair:
          type: string
        direction:
          type: string
        timeframe:
          type: string
        overallScore:
          type: integer
        scoreLabel:
          type: string
        verdict:
          type: string
        outcome:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RiskCalcInput:
      type: object
      required: [accountSize, riskPercent, pair, entryPrice, stopLoss, takeProfit]
      properties:
        accountSize:
          type: number
        riskPercent:
          type: number
        pair:
          type: string
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        leverage:
          type: ["number", "null"]
    RiskCalcResult:
      type: object
      required: [riskDollars, rewardDollars, rrRatio, lotSize, stopPips, takeProfitPips, pipValue, marginRequired, maxLossScenarios]
      properties:
        riskDollars:
          type: number
        rewardDollars:
          type: number
        rrRatio:
          type: number
        lotSize:
          type: number
        stopPips:
          type: number
        takeProfitPips:
          type: number
        pipValue:
          type: number
        marginRequired:
          type: ["number", "null"]
        breakEvenPips:
          type: number
        maxLossScenarios:
          type: array
          items:
            type: object
    RiskScenario:
      type: object
      required: [id, userId, name, accountSize, riskPercent, pair, entryPrice, stopLoss, takeProfit, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        accountSize:
          type: number
        riskPercent:
          type: number
        pair:
          type: string
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: ["number", "null"]
        pipValue:
          type: ["number", "null"]
        riskDollars:
          type: ["number", "null"]
        rewardDollars:
          type: ["number", "null"]
        rrRatio:
          type: ["number", "null"]
        createdAt:
          type: string
    RiskScenarioInput:
      type: object
      required: [name, accountSize, riskPercent, pair, entryPrice, stopLoss, takeProfit]
      properties:
        name:
          type: string
        accountSize:
          type: number
        riskPercent:
          type: number
        pair:
          type: string
        entryPrice:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        lotSize:
          type: ["number", "null"]
        pipValue:
          type: ["number", "null"]
        riskDollars:
          type: ["number", "null"]
        rewardDollars:
          type: ["number", "null"]
        rrRatio:
          type: ["number", "null"]