# lib/api-zod/src/generated/types/weeklyReport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TradeScope API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyReport {
  summary: string;
  strengths: string[];
  improvements: string[];
  tip: string;
  generatedAt: string;
}